requirements.txt
python-dotenv==1.0.0
openai
PyPDF2==3.0.1 